requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
jsonschema>=4.17.0
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.35.0
joblib>=1.3.0
ta>=0.11.0
requests>=2.31.0
watchdog>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
